realign egui context into View struct
