+++ src/ui/view.rs
@@ -1,9 +1,19 @@
+use eframe::egui;
+use crate::io::{Reader, Writer};
+
 pub struct View {
     infile: Option<String>,
     outfile: Option<String>
 }
 
 impl View {
+    pub fn new() -> Self {
+        View {
+            infile: None,
+            outfile: None
+        }
+    }
+
     pub fn get_infile_name(&self) -> Result<String, std::io::Error> {
         match &self.infile {
             Some(s) => Ok(s.clone()),
@@ -18,5 +28,68 @@ impl View {
             None => Err(std::io::Error::new(std::io::ErrorKind::NotFound,
                                             "You must select an output file!"))
         }
+    }
+}
+
+impl eframe::App for View {
+    fn update(&mut self, ctx: &egui::Context, _frame: &mut eframe::Frame) {
+        egui::CentralPanel::default().show(ctx, |ui| {
+            ui.label("Select an input file:");
+            if ui.button("Select input file…").clicked() {
+                if let Some(path) = rfd::FileDialog::new().pick_file() {
+                    self.input_filename = Some(path.display().to_string());
+                }
+            }
+            if let Some(picked_path) = &self.input_filename {
+                ui.horizontal(|ui| {
+                    ui.label("Picked file:");
+                    ui.monospace(picked_path);
+                });
+            }
+
+            ui.label("Select an output file:");
+            if ui.button("Select output file…").clicked() {
+                if let Some(path) = rfd::FileDialog::new().pick_file() {
+                    self.output_filename = Some(path.display().to_string());
+                }
+            }
+            if let Some(picked_path) = &self.output_filename {
+                ui.horizontal(|ui| {
+                    ui.label("Picked file:");
+                    ui.monospace(picked_path);
+                });
+            }
+        });
+    }
+}
+
+/// Preview hovering files:
+fn _preview_files_being_dropped(ctx: &egui::Context) {
+    use egui::*;
+
+    if !ctx.input().raw.hovered_files.is_empty() {
+        let mut text = "Dropping files:\n".to_owned();
+        for file in &ctx.input().raw.hovered_files {
+            if let Some(path) = &file.path {
+                text += &format!("\n{}", path.display());
+            } else if !file.mime.is_empty() {
+                text += &format!("\n{}", file.mime);
+            } else {
+                text += "\n???";
+            }
+        }
+
+        let painter =
+            ctx.layer_painter(LayerId::new(Order::Foreground, Id::new("file_drop_target")));
+
+        let screen_rect = ctx.input().screen_rect();
+        painter.rect_filled(screen_rect, 0.0, Color32::from_black_alpha(192));
+        painter.text(
+            screen_rect.center(),
+            Align2::CENTER_CENTER,
+            text,
+            TextStyle::Heading.resolve(&ctx.style()),
+            Color32::WHITE,
+        );
     }
 }
